added method to find All products that user hasn't evaluated yet, and method to see all users products with avg evaluation and count of comments

+++ src/main/java/com/jdbc/hacaton1/controllers/EvaluationController.java
@@ -18,7 +18,7 @@ public class EvaluationController {
     }
 
     @GetMapping("getAvgEvaluationByProductId:{id}")
-        public Integer getAvgEvaluation(@PathVariable Integer id){
+        public Double getAvgEvaluation(@PathVariable Integer id){
         return service.getAvgEvaluation(id);
     }
 }
